Extract fellowship salary claimants query (#4088)
