+++ packages/next-common/hooks/fellowship/salary/useFellowshipSalaryClaimantsData.js
@@ -1,42 +1,17 @@
 import { find, orderBy } from "lodash-es";
-import { useContextApi } from "next-common/context/api";
 import { usePageProps } from "next-common/context/page";
-import {
-  fellowshipSalaryClaimantsSelector,
-  fellowshipSalaryClaimantsTriggerUpdateSelector,
-  fetchFellowshipSalaryClaimants,
-  setFellowshipSalaryClaimants,
-} from "next-common/store/reducers/fellowship/claimants";
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { fellowshipSalaryClaimantsSelector } from "next-common/store/reducers/fellowship/claimants";
+import { useSelector } from "react-redux";
 
 export function useFellowSalaryClaimantsData() {
-  const { fellowshipSalaryClaimants = [], fellowshipMembers = [] } =
-    usePageProps();
-  const api = useContextApi();
-  const dispatch = useDispatch();
+  const {
+    fellowshipSalaryClaimants: claimantsFromServer = [],
+    fellowshipMembers = [],
+  } = usePageProps();
+  const claimantsFromRedux = useSelector(fellowshipSalaryClaimantsSelector);
+  const dataSource = claimantsFromRedux || claimantsFromServer;
 
-  const claimants = useSelector(fellowshipSalaryClaimantsSelector);
-  const triggerUpdate = useSelector(
-    fellowshipSalaryClaimantsTriggerUpdateSelector,
-  );
-
-  const isFirst = triggerUpdate === 0;
-  const dataSource = isFirst ? fellowshipSalaryClaimants : claimants;
-
-  useEffect(() => {
-    dispatch(setFellowshipSalaryClaimants(fellowshipSalaryClaimants));
-  }, [dispatch, fellowshipSalaryClaimants]);
-
-  useEffect(() => {
-    if (!api) {
-      return;
-    }
-
-    dispatch(fetchFellowshipSalaryClaimants(api));
-  }, [api, triggerUpdate]);
-
-  const data = orderBy(
+  return orderBy(
     dataSource.map((claimant) => {
       const address = claimant?.address;
       const member = find(fellowshipMembers, { address });
@@ -50,6 +25,4 @@ export function useFellowSalaryClaimantsData() {
     "rank",
     "desc",
   );
-
-  return data;
 }
